chore: updated the logic and added sync tasks in celery imports
