+++ apiserver/plane/db/management/commands/sync_issue_version.py
@@ -2,7 +2,7 @@
 from django.core.management.base import BaseCommand
 
 # Module imports
-from plane.bgtasks.issue_version_task import schedule_issue_management_command_task
+from plane.bgtasks.issue_version_sync import schedule_issue_version
 
 
 class Command(BaseCommand):
@@ -12,7 +12,7 @@ class Command(BaseCommand):
         batch_size = input("Enter the batch size: ")
         batch_countdown = input("Enter the batch countdown: ")
 
-        schedule_issue_management_command_task.delay(
+        schedule_issue_version.delay(
             batch_size=int(batch_size), countdown=int(batch_countdown)
         )
